Separação do sistema de desenho do sistema de infos. de debug. Alteração na chamada do desenho de sprites. Suporte a sprites com uma só TextureRegion

+++ core/src/com/stephenfg/sre/componentes/ComponenteTransformacao.java
@@ -7,7 +7,7 @@ public class ComponenteTransformacao implements Component {
     public Vector2 posicao;
     public Vector2 escala;
 
-    public double rotacao = 0.f;
+    public float rotacao = 0.f;
 
     public ComponenteTransformacao(Vector2 posicao){
         this.posicao = posicao;
